+++ dxsp/handler/oneinch.py
@@ -2,7 +2,7 @@
 # OneInch 🦄
 # """
 
-# from .client import DexClient
+# from ._client import DexClient
 
 
 # class OneinchHandler(DexClient):
